pyproject.toml
[project]
name = "lms-backend"
version = "0.1.0"
description = "FastAPI backend for LMS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "email-validator>=2.1.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9"
  ,"bcrypt<4"
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0"
]
